# docs/openapi_skeleton.yaml
openapi: 3.1.0
info:
  title: Gestionale Collaboratori e Progetti
  description: "\n    Sistema gestionale completo per:\n    - \U0001F465 Collaboratori\
    \ (anagrafica, ruoli, tariffe)\n    - \U0001F4CB Progetti (budget, ente attuatore,\
    \ stato)\n    - \U0001F4DD Incarichi (assegnazioni ai progetti con mansioni)\n\
    \    - ⏰ Presenze (rilevazioni orarie con validazioni)\n    - \U0001F4C4 Contratti\
    \ (collaborazione, P.IVA, subordinato)\n    - \U0001F4B0 Rendicontazione (FSE+,\
    \ Fondimpresa, PNRR)\n    "
  version: 1.0.0
paths:
...
  /api/v1/projects/:
    post:
      tags:
      - Projects
      summary: Create Project
      description: Crea un nuovo progetto
      operationId: create_project_api_v1_projects__post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Projects
      summary: Get Projects
      description: Lista tutti i progetti con filtro status opzionale
      operationId: get_projects_api_v1_projects__get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
                title: Response Get Projects Api V1 Projects  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}:
    get:
      tags:
      - Projects
      summary: Get Project
      description: Ottieni un progetto specifico
      operationId: get_project_api_v1_projects__project_id__get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Projects
      summary: Update Project
      description: Aggiorna un progetto
      operationId: update_project_api_v1_projects__project_id__put
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Projects
      summary: Delete Project
      description: Elimina un progetto
      operationId: delete_project_api_v1_projects__project_id__delete
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/entities/:
    post:
      tags:
      - Entities
      summary: Create Entity
      description: Crea un nuovo ente attuatore
      operationId: create_entity_api_v1_entities__post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Entities
      summary: Get Entities
      description: Lista tutti gli enti attuatori con ricerca opzionale
      operationId: get_entities_api_v1_entities__get
      parameters:
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entity'
                title: Response Get Entities Api V1 Entities  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/entities/{entity_id}:
    get:
      tags:
      - Entities
      summary: Get Entity
      description: Ottieni un ente attuatore specifico
      operationId: get_entity_api_v1_entities__entity_id__get
      parameters:
      - name: entity_id
        in: path
        required: true
        schema:
          type: integer
          title: Entity Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Entities
      summary: Update Entity
      description: Aggiorna un ente attuatore
      operationId: update_entity_api_v1_entities__entity_id__put
      parameters:
      - name: entity_id
        in: path
        required: true
        schema:
          type: integer
          title: Entity Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Entities
      summary: Delete Entity
      description: Elimina un ente attuatore
      operationId: delete_entity_api_v1_entities__entity_id__delete
      parameters:
      - name: entity_id
        in: path
        required: true
        schema:
          type: integer
          title: Entity Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/reporting/timesheet:
    get:
      tags:
      - Reporting
      summary: Get Timesheet
      description: Report timesheet presenze
      operationId: get_timesheet_api_v1_reporting_timesheet_get
      parameters:
      - name: from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: From
      - name: to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: To
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reporting/summary:
    get:
      tags:
      - Reporting
      summary: Get Summary
      description: Report riepilogativo KPI
      operationId: get_summary_api_v1_reporting_summary_get
      parameters:
      - name: from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: From
      - name: to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: To
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Entity:
      properties:
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        vat_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Vat Number
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - name
      - id
      - created_at
      title: Entity
    EntityCreate:
      properties:
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        vat_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Vat Number
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
      type: object
      required:
      - name
      title: EntityCreate
    EntityUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        vat_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Vat Number
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
      type: object
      title: EntityUpdate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Project:
      properties:
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start Date
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
        status:
          type: string
          title: Status
          default: active
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - name
      - id
      - created_at
      title: Project
    ProjectCreate:
      properties:
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start Date
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
        status:
          type: string
          title: Status
          default: active
      type: object
      required:
      - name
      title: ProjectCreate
    ProjectUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start Date
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      type: object
      title: ProjectUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError